move epxr::binary
