+++ datafusion/physical-expr-common/Cargo.toml
@@ -39,3 +39,4 @@ path = "src/lib.rs"
 arrow = { workspace = true }
 datafusion-common = { workspace = true, default-features = true }
 datafusion-expr = { workspace = true }
+paste = "^1.0"
