+++ datafusion/physical-expr-common/src/lib.rs
@@ -18,6 +18,7 @@
 pub mod aggregate;
 pub mod expressions;
 pub mod physical_expr;
+pub mod intervals;
 pub mod sort_expr;
 pub mod sort_properties;
 pub mod tree_node;
